Break out Fedora 3 functions
This commit breaks the Fedora 3 functions into their own folder
so that we can go about implementing other repositories in
tuque.
