+++ implementations/fedora3/FedoraApiSerializer.php
