+++ implementations/fedora3/FoxmlDocument.php
